requirements.txt
Django>=5.2.1
djangorestframework
django-filter
drf-yasg 